+++ contracts/core/defaults/WitnetRequestBoardTrustableDefault.sol
@@ -73,7 +73,7 @@ contract WitnetRequestBoardTrustableDefault
     /// @notice Estimate the minimum reward required for posting a data request with a callback.
     /// @param _gasPrice Expected gas price to pay upon posting the data request.
     /// @param _callbackGasLimit Maximum gas to be spent when reporting the data request result.
-    function estimateBaseFeeWithCallback(uint256 _gasPrice, uint96 _callbackGasLimit)
+    function estimateBaseFeeWithCallback(uint256 _gasPrice, uint24 _callbackGasLimit)
         public view
         virtual override
         returns (uint256)
